requirements/requirements_dup.txt
pandas>=2.0
numpy>=1.24
Pillow>=10.0
ImageHash>=4.3.1
tqdm>=4.66